# Allow to actually search against new search index