This is a followup to #831 which adds logic to perform project search queries against the external MeiliSearch index instead of using Postgres search. 

When results are received in the shape of permalinks from the search index, subsequently the corresponding projects are fetched once again from postgres, which in my tests was much much quicker.

As I'm still experimenting with this and I'd like to verify that my prior performance tests actually hold true this ships this feature behind a feature flag to enable some additional testing of this before I then hopefully move on to fully replace the existing search code with this (and throwing in some additional specs and refactoring the not-yet-optimal order query stuff)