@colszowka 2021-02-27
Ok, looks like a decent improvement so far:

```
$ ab -n 10 'https://www.ruby-toolbox.com/search?q=code+coverage'

# OLD
Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:      116  130   7.1    133     137
Processing:  8680 12187 1416.3  12429   13582
Waiting:     8566 12067 1413.1  12310   13466
Total:       8802 12316 1418.9  12546   13718

# NEW
Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:      115  131   8.4    135     141
Processing:   781  971 149.9   1012    1195
Waiting:      666  858 149.2    887    1081
Total:        917 1102 150.2   1147    1330
```

Problems that I've found:

* When I enable the show bugfix forks toggle the request times out, this needs some further investigation on what goes wrong there
* When more projects are actually returned from the search index, things are becoming slower. Since the search rank order sql stuff always lists the entire search result maybe this slows postgres down. Maybe this could be optimized somehow? Also, maybe the limit of 1000 search results is too high as well, not sure if anybody browses to page 30 really..