@Freaky 2021-03-20
Was any effort made to examine the Postgres query for this?  You're doing full text search over

```
 select sum(length(description) + length(permalink)) from projects;
   sum
----------
 15786381
```

A whopping 16MB of data.  You don't need some newfangled hyper-optimized Rust search engine for that, you just need a query plan that isn't bonkers.  Postgres is _more_ than capable:

```
rubytoolbox=# SELECT permalink, ts_rank_cd(description_tsvector, query)*(score+1) AS rank
  FROM projects, to_tsquery('simple','code & coverage') query
  WHERE query @@ description_tsvector ORDER BY rank DESC;
           permalink           |        rank
-------------------------------+---------------------
 simplecov                     |    1.40300002090633
 simplecov-html                |    1.15900001727045
 chefspec                      |   0.418000006228685
...
(83 rows)

Time: 2.480 ms
```