@colszowka 2021-03-28
@Freaky Thanks for your input. First of all let me say that I would really appreciate if we could keep the tone here a bit less confrontational. You can rest assured that I value my own time greatly and would not do a switch like this just because it feels more shiny or for fun. So to answer your question - yes, of course I tried to examine what might be wrong first.

So, why did I make this switch?

* When I [originally implemented the search feature](https://github.com/rubytoolbox/rubytoolbox/pull/84) I specifically chose to go with pg_search because I too deemed it absolutely good enough and wanted to avoid having an extra dependency to deal with and run. I also tried following the [advice from the pg_search](https://github.com/Casecommons/pg_search#using-tsvector-columns) gem to optimize it. The speeds were not amazing but reasonable for quite some time
* At some point the performance started to degrade on the production site. I cannot really pinpoint this to when though. I am not aware of making any changes to the queries etc. in recent time, however a while ago the [addition of historical gem download data](https://www.ruby-toolbox.com/blog/2019-02-25/historical-gem-download-charts) increased the overall database size significantly
* So keeping in mind that I am not aware of any real changes to the database queries, request volume etc, the performance being just absolutely terrible and pg explain not giving me much hope that I could optimize this I decided to run a little experiment using another indexing solution, which proved so much more performant that all things considered my impression was I could provide a better search experience to Toolbox users more quickly and with less effort by just swapping the indexing approach.

As mentioned above, I personally would absolutely prefer to use pg_search and avoid having another dependency, and all code for indexing against pg_search is still in place. So if you are more familiar than me with optimizing pg_search I'd appreciate your input here as to how to make it work better, the [current data base dumps are as usual available here](https://data.ruby-toolbox.com/), and on production the site is running on a Heroku standard-0 postgres plan